Moving concat to utils
